refactor(review-providers): extract shared parseProviderFindings parser (refs #3981)
